+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[BREAKING CHANGE]** Added an `open` & `innerClassName` props on `<Drawer>`
 - **[FIX]** Remove wrong coloring on AutoComplete when using keyboard to navigate
 
 # v0.10.2 (10/09/2018)